% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConnectivity.R
\name{checkConnectivity}
\alias{checkConnectivity}
\title{Check the connectivity of parcels in a parcellation}
\usage{
checkConnectivity(parcellation, parcel = NA,
  shape.mat = neighborShape_Box27())
}
\arguments{
\item{parcellation}{a \code{BrcParcellation} object}

\item{parcel}{a vector of indicies of which parcels to check the connectivity of}

\item{shape.mat}{the output of \code{neighborShape} function, such as \code{neighborShape_Box27()}}
}
\value{
a vector of booleans
}
\description{
This function checks to see if each parcel in a parcellation are connected.
This can be useful is one wants ensure that no parcel consists of two
disjoint sets of voxels.
}
\details{
By default, \code{parcel} is set to \code{NA}, meaning this function will check
every parcel.

This function returns a vector of \code{TRUE}, \code{FALSE} values, one for
each parcel.
}
