% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMask.R
\name{createMask}
\alias{createMask}
\title{Create a mask from an existing parcellation}
\usage{
createMask(parcellation)
}
\arguments{
\item{parcellation}{a \code{BrcParcellation} object}
}
\value{
a \code{BrcParcellation} object
}
\description{
This function takes in a \code{BrcParcellation} object and returns
a new \code{BrcParcellation} object where all the voxels assigned
to any parcels in the input parcellation are now assigned to parcel 1.
}
\details{
This is useful if the user wants to make a mask, so downstream, the user
can zero-out all the voxels not in the parcellation for a region-of-interest
(RoI) study.
}
