% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceBetween.R
\name{distance_manhattan}
\alias{distance_manhattan}
\title{A distance function between to matrices (Manhattan distance)}
\usage{
distance_manhattan(mat1, mat2)
}
\arguments{
\item{mat1}{a matrix}

\item{mat2}{a matrix}
}
\value{
a numeric
}
\description{
The input to this function are two matrices with the same number of rows.
This function computes the minimum
distance between any column in \code{mat1} to any column in \code{mat2}.
}
