% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceBetween.R
\name{distanceBetween}
\alias{distanceBetween}
\title{Compute the distance between parcels in a parcellation}
\usage{
distanceBetween(parcellation, dist = distance_manhattan, ...)
}
\arguments{
\item{parcellation}{a \code{BrcParcellation} object}

\item{dist}{a distance function}

\item{...}{additional arguments to be passed into \code{dist} (optional)}
}
\value{
a \code{dist} object
}
\description{
This function computes the distance between parcels in a given parcellation.
}
\details{
The distance between two parcels is passed into this function as another function,
such as \code{distance_manhattan} for example, as the \code{dist} argument.
The \code{dist} function must be able to take two matrices, each with 
3 rows, denoting the 3-dimensional indices (locations) of all the voxels in
each of the two parcels.
}
