% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillParcellation.R
\name{fillParcellation}
\alias{fillParcellation}
\title{Fills all the empty voxels in a parcellation}
\usage{
fillParcellation(parcellation, shape.mat = neighborShape_Box27())
}
\arguments{
\item{parcellation}{a \code{BrcParcellation} object}

\item{shape.mat}{the output of \code{neighborShape} function, such as \code{neighborShape_Box27()}}
}
\value{
a \code{BrcParcellation} object
}
\description{
If a parcellation (\code{BrcParcellation} object) has empty voxels,
this function returns a new parcellation such that every empty voxel is
assigned to same parcel as its nearest non-empty voxel. If an empty 
voxel is equidistant to more than one non-empty nearest voxel that differ
in parcel assignments, the empty voxel is randomly assigned to one of the
possible parcels.
}
\details{
As all parcellations represent 3-dimensional objects (given by
\code{parcellation$dim3d}), this function will assign all the empty voxel
in the 3-dimensional space to a parcel.
}
