% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammingDistance.R
\name{hammingDistance}
\alias{hammingDistance}
\title{Compute the pairwise distance between two parcellations: Pairwise hamming distance}
\source{
Ben-David, Shai, Ulrike Von Luxburg, and Dávid Pál. 
"A sober look at clustering stability." International Conference on 
Computational Learning Theory. Springer, Berlin, Heidelberg, 2006, (Definition 4)
}
\usage{
hammingDistance(parcellation1, parcellation2, numPairs = 5000)
}
\arguments{
\item{parcellation1}{a \code{BrcParcellation} object}

\item{parcellation2}{a \code{BrcParcellation} object}

\item{numPairs}{number of pairs of voxels to compute the pairwise hamming distance over}
}
\value{
a numeric
}
\description{
Given two \code{BrcParcellation} objects, \code{parcellation1} and
\code{parcellation2}, compute the pairwise Hamming distance. This is defined the 
following way: for random pairs of voxels (which are assigned to voxels in
both parcellations), count a distance of 0 if both voxels are either in the same
parcel within each parcellation or if both voxels are in different parcels 
within each parcellation, but count a distance of 1 if both voxels are in the
same parcel in one parcellation but in different parcels in the other parcellation.
The distance function then is the average of these (0,1) values over all the
pairs.
}
\details{
This distance is more accurate when \code{numPairs} is larger, at the cost
of more computational time.
}
