% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectParcellation.R
\name{intersectParcellation}
\alias{intersectParcellation}
\title{Intersect two parcellations}
\usage{
intersectParcellation(parcellation1, parcellation2)
}
\arguments{
\item{parcellation1}{a \code{BrcParcellation} object}

\item{parcellation2}{a \code{BrcParcellation} object}
}
\value{
a \code{BrcParcellation} object
}
\description{
This function creates a new parcellation from two existing parcellations
where voxels in the new parcellation are in the same parcel if and only if
the voxels were in the same parcels within each of the two input parcellations.
}
