% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborOthers.R
\name{neighborParcel2Parcel}
\alias{neighborParcel2Parcel}
\title{Find neighboring parcels for each parcel}
\usage{
neighborParcel2Parcel(parcellation, parcel = NA,
  shape.mat = neighborShape_Box27())
}
\arguments{
\item{parcellation}{a \code{BrcParcellation} object}

\item{parcel}{a vector of indicies of which parcels to select the neighbors of}

\item{shape.mat}{the output of \code{neighborShape} function, such as \code{neighborShape_Box27()}}
}
\value{
A list of integers, one vector per parcel
}
\description{
Given a \code{BrcParcellation} object and a vector of parcel
indicies in \code{parcel}, find all parcels that neighbor each of the
\code{parcels}.By default, \code{parcel} is set to \code{NA}, meaning
it finds the neighbor of all parcels.
}
\details{
The \code{parcel} vector must be positive integers ranging between
\code{1} and \code{max(parcellation$partition)}.
}
