% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborOthers.R
\name{neighborParcel2Voxel}
\alias{neighborParcel2Voxel}
\title{Find neighboring voxels for each parcel}
\usage{
neighborParcel2Voxel(parcellation, parcel = NA,
  shape.mat = neighborShape_Box27())
}
\arguments{
\item{parcellation}{a \code{BrcParcellation} object}

\item{parcel}{a vector of indicies of which parcels to select the neighbors of}

\item{shape.mat}{the output of \code{neighborShape} function, such as \code{neighborShape_Box27()}}
}
\value{
A list of integers, one vector per parcel
}
\description{
Given a \code{BrcParcellation} object and a vector of parcel
indicies in \code{parcel}, find all voxels that neighbor each of the
\code{parcel}.By default, \code{parcel} is set to \code{NA}, meaning
it finds the neighbor of all parcel
}
\details{
The \code{parcel} vector must be positive integers ranging between
\code{1} and \code{max(parcellation$partition)}.

The voxel vector are indicies referring of specific locations in
\code{parcellation}. They are positive integers between \code{1} and 
\code{length(parcellation$partition)}.
}
