% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborVoxel2Voxel.R
\name{neighborShape_Box27}
\alias{neighborShape_Box27}
\title{Neighbor shape for box}
\usage{
neighborShape_Box27()
}
\value{
a matrix with 3 columns
}
\description{
This is a function needed for many functions such as 
\code{neighborVoxel2Voxel} to determine how neighbors are defined.
This will typically be the \code{shape.mat} argument to many such functions.
}
\details{
This shape is defined in such a way that if the voxel in question is
the middle voxel of a 3x3x3 cube, all the voxels consisting of the 3x3x3
cube are the neighboring voxels.
}
