% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborVoxel2Voxel.R
\name{neighborShape_Diamond7}
\alias{neighborShape_Diamond7}
\title{Neighbor shape for diamond}
\usage{
neighborShape_Diamond7()
}
\value{
a matrix with 3 columns
}
\description{
This is a function needed for many functions such as 
\code{neighborVoxel2Voxel} to determine how neighbors are defined.
This will typically be the \code{shape.mat} argument to many such functions.
}
\details{
This shape is defined in such a way that if the voxel in question is
a 6-sided die, the neighboring voxels are the 6 voxels immediately bordering
each face of the die.
}
