% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborVoxel2Voxel.R
\name{neighborVoxel2Voxel}
\alias{neighborVoxel2Voxel}
\title{Find neighboring voxels for each voxel}
\usage{
neighborVoxel2Voxel(parcellation, voxel = NA,
  shape.mat = neighborShape_Box27())
}
\arguments{
\item{parcellation}{a \code{BrcParcellation} object}

\item{voxel}{a vector of indicies of which voxels to select the neighbors of}

\item{shape.mat}{the output of \code{neighborShape} function, such as \code{neighborShape_Box27()}}
}
\value{
A list of integers, one vector per voxel
}
\description{
Given a \code{BrcParcellation} object and a vector of voxels
indices in \code{voxel}, find all the voxels that neighbor each of
the voxels. By default, \code{voxel} is set to \code{NA}, meaning
it finds the neighbor of all voxels.
}
\details{
The \code{voxel} vector are indicies referring of specific locations in
\code{parcellation}. They must be positive integers between \code{1} and 
\code{length(parcellation$partition)}.
}
