% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentMaskOverlap.R
\name{percentMaskOverlap}
\alias{percentMaskOverlap}
\title{Compute the percentage of overlap of non-empty voxels between two parcellations}
\usage{
percentMaskOverlap(parcellation1, parcellation2)
}
\arguments{
\item{parcellation1}{a \code{BrcParcellation} object}

\item{parcellation2}{a \code{BrcParcellation} object}
}
\value{
a numeric
}
\description{
This function takes in two parcellations and computes the following the
fraction: the numerator represents the number of voxels that are non-empty in
both parcellations, and the denominator represents the number of voxels are 
non-empty in either parcellations.
}
