% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceArea.R
\name{surfaceArea}
\alias{surfaceArea}
\title{Compute the surface area of each parcel}
\usage{
surfaceArea(parcellation, parcel = NA, shape.mat = neighborShape_Box27())
}
\arguments{
\item{parcellation}{a \code{BrcParcellation} object}

\item{parcel}{a vector of indicies of which parcels to compute the surface area for}

\item{shape.mat}{the output of \code{neighborShape} function, such as \code{neighborShape_Box27()}}
}
\value{
a vector of numerics
}
\description{
This function computes the surface area for each parcel (i.e.,
how many voxels are on the surface of a parcel) and returns it as a 
vector, one number for each parcel.
}
\details{
The definition of which voxel lie
on the surface is dictated by \code{shape.mat}, where a voxel is 
on the surface if one of its neighbors belong to a different parcel or
one of its neighbor is an empty voxel.
}
